AdvLoggerPkg: Fix memory bucket stability
Right now, the PEI Core instance of AdvancedLoggerLib will always
allocate the logger buffer as `EfiRuntimeServicesData`. This means
it is not allocated into the DXE Core managed RT Services Data
memory bucket. The DXE Core instance of AdvancedLoggerLib either
continues to use this buffer it is provided or allocates a new
`EfiReservedMemoryType` buffer.

This change always allocates a `EfiReservedMemoryType` buffer in
the DXE Core instance to ensure it is allocated from the reserved
memory type buckets setup by the DXE Core. The PEI Core logger
buffer is updated to be `EfiBootServicesData` so it does not affect
runtime allocations.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>